add support for using jackc/pgx as a driver
This patch updates `pggen` to support using `jackc/pgx` as
a database driver. `lib/pq` has entered maintenance mode, and
`jackc/pgx` is the recommended replacement. I still cannot
recommend using `jackc/pgx` as a driver at this time because it
has [a bug][1] that prevents `pggen` from gracefully handling
schema changes.

Most of the changes involve tweaking our scanners to be able
to consume both the data that `lib/pq` and `jackc/pgx` pass
down.

For the moment, the generated code still depends on `lib/pq`
because we need to be able to use `lib/pq.Array` to scan arrays
of values. Once `jackc/pgx` becomes the recommended driver, we
will switch to using whatever the equivalent of `lib/pq.Array`
is for `jackc/pgx` (or roll our own). We will not drop support
for using `lib/pq` as a driver (likely not ever, but certainly
not without a major version bump).

This bumps our go versions, including the msgv. We need to bump
the msgv because jackc/pgx does not support go 1.11.

Closes #119

[1]: https://github.com/jackc/pgx/issues/841

+++ cmd/pggen/test/tables_test.go
@@ -2,6 +2,7 @@ package test
 
 import (
 	"database/sql"
+	"os"
 	"reflect"
 	"sort"
 	"testing"
@@ -1118,6 +1119,12 @@ func TestNotFoundList(t *testing.T) {
 }
 
 func TestDroppingColumnOnTheFly(t *testing.T) {
+	if os.Getenv("DB_DRIVER") == "pgx" {
+		// disable this test for jackc/pgx
+		// TODO: re-enable once https://github.com/jackc/pgx/issues/841 is resolved
+		return
+	}
+
 	// make sure we always start in a consistant state
 	_, err := pgClient.Handle().ExecContext(ctx, `
 		DROP TABLE drop_cols;
@@ -1144,6 +1151,8 @@ func TestDroppingColumnOnTheFly(t *testing.T) {
 
 	// load the record again
 	dc, err := pgClient.GetDropCol(ctx, id)
+	// NOTE: error actually occurs in github.com/jackc/pgx/v4/stdlib.(*Conn).QueryContex
+	//       when scanning the very first row.
 	chkErr(t, err)
 
 	if dc.F1 != 0 {
@@ -1154,7 +1163,7 @@ func TestDroppingColumnOnTheFly(t *testing.T) {
 		t.Fatalf("expected F2 to be 2, was %d", dc.F2)
 	}
 
-	// pull the rug out from under us
+	// replace the old column
 	_, err = pgClient.Handle().ExecContext(ctx, `ALTER TABLE drop_cols ADD COLUMN f1 int NOT NULL DEFAULT 1`)
 	chkErr(t, err)
 }
